streamlit-service/requirements.txt
opencv-python
pillow
python-io
requests
numpy
streamlit==1.40.1